animal-hotel-rest/src/main/java/ru/geekbrains/service/RoomService.java
package ru.geekbrains.service;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import ru.geekbrains.entity.Room;
import ru.geekbrains.repository.RoomRepository;

import java.util.List;
import java.util.Optional;

@Service
@RequiredArgsConstructor
public class RoomService {

    private final RoomRepository roomRepository;

    public List<Room> findAll() {
        return roomRepository.findAll();
    }

    public Optional<Room> findById(Long id) {
        return roomRepository.findById(id);
    }

    public Room save(Room room) {
        return roomRepository.save(room);
    }

    public void deleteById(Long id) {
        roomRepository.deleteById(id);
    }

    public List<Room> findByHotelId(Long id) {
        return roomRepository.findByHotelId(id);
    }

    public List<Room> findByRoomTypeId(Long id) {
        return roomRepository.findByRoomTypeId(id);
    }
}
